requirements.txt
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
scipy==1.1.0
six==1.11.0
